refactor instructions to separate file
