+++ src/lc3/mod.rs
@@ -1,4 +1,5 @@
 pub mod bytes;
+pub mod instructions;
 pub mod memory;
 pub mod opcode;
 pub mod vm;
